Distribute Texter on Player and Narrator
The classes lost their common basis due to the need for letter by
letter changes which didn't seem feasable for the player.
